requirements.txt
# TCPD - Tester's Comprehensive PC Diagnostics - Dependencies
# Install with: pip install -r requirements.txt

# System Information
psutil>=5.9.0
wmi>=1.5.1
pywin32>=306
py-cpuinfo>=9.0.0

# GPU Monitoring
pynvml>=11.5.0
GPUtil>=1.4.0

# GPU Compute (Stress Testing)
pyopencl>=2023.1.0
numpy>=1.24.0

# Storage Health
pySMART>=1.3.0

# CLI Framework
typer[all]>=0.9.0
rich>=13.7.0
questionary>=2.0.0     # Arrow-key TUI menus

# Configuration
pyyaml>=6.0

# Data Models
pydantic>=2.0

# Build (dev only)
pyinstaller>=6.0
pyarmor>=8.0           # Code obfuscation for protected builds
